requirements.txt
streamlit
flask
flask_sqlalchemy
pandas
matplotlib
